New major changes
Instead of uploading to DHIS2 all the time we've added local json
objects to test out the code. This way we'll save a lot of time in the
future. After planning with Thuc over the weekend we've decided to go
back to selector instead of using a button (to retrieve data easier).

Cleaning up code in dataFetching.js and the pages.